# Optimizations to aquifers, noise sampling, and random number generation