--- a/src/main/java/io/github/opencubicchunks/cubicchunks/mixin/core/common/chunk/MixinDepthBasedReplacingBaseStoneSource.java
+++ b/src/main/java/io/github/opencubicchunks/cubicchunks/mixin/core/common/chunk/MixinDepthBasedReplacingBaseStoneSource.java
@@ -7,6 +7,7 @@ import net.minecraft.world.level.levelgen.NoiseGeneratorSettings;
 import net.minecraft.world.level.levelgen.WorldgenRandom;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.Mutable;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -16,13 +17,11 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 @Mixin(DepthBasedReplacingBaseStoneSource.class)
 public class MixinDepthBasedReplacingBaseStoneSource {
 
+    @Shadow @Final private BlockState normalBlock;
     @Shadow @Final private BlockState replacementBlock;
 
-    @Shadow @Final private WorldgenRandom random;
-
-    @Shadow @Final private BlockState normalBlock;
-
     @Shadow @Final private long seed;
+    @Mutable @Shadow @Final private WorldgenRandom random;
 
     private long seedX, seedY, seedZ;