--- a/src/main/java/io/github/opencubicchunks/cubicchunks/mixin/core/common/chunk/MixinDepthBasedReplacingBaseStoneSource.java
+++ b/src/main/java/io/github/opencubicchunks/cubicchunks/mixin/core/common/chunk/MixinDepthBasedReplacingBaseStoneSource.java
@@ -1,5 +1,6 @@
 package io.github.opencubicchunks.cubicchunks.mixin.core.common.chunk;
 
+import io.github.opencubicchunks.cubicchunks.chunk.NonAtomicWorldgenRandom;
 import net.minecraft.util.Mth;
 import net.minecraft.world.level.block.state.BlockState;
 import net.minecraft.world.level.levelgen.DepthBasedReplacingBaseStoneSource;
@@ -7,7 +8,6 @@ import net.minecraft.world.level.levelgen.NoiseGeneratorSettings;
 import net.minecraft.world.level.levelgen.WorldgenRandom;
 import org.spongepowered.asm.mixin.Final;
 import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.Mutable;
 import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
@@ -21,7 +21,7 @@ public class MixinDepthBasedReplacingBaseStoneSource {
     @Shadow @Final private BlockState replacementBlock;
 
     @Shadow @Final private long seed;
-    @Mutable @Shadow @Final private WorldgenRandom random;
+    @Shadow @Final private WorldgenRandom random;
 
     private long seedX, seedY, seedZ;
 
@@ -44,9 +44,9 @@ public class MixinDepthBasedReplacingBaseStoneSource {
         } else if (probability >= 1) {
             cir.setReturnValue(this.replacementBlock);
         } else {
-            long seed = x * this.seedX ^ y * this.seedY ^ z * this.seedZ ^ this.seed;
-            this.random.setSeed(seed);
-            cir.setReturnValue(this.random.nextFloat() < probability ? this.replacementBlock : this.normalBlock);
+            long seed = NonAtomicWorldgenRandom.next(x * this.seedX ^ y * this.seedY ^ z * this.seedZ ^ this.seed);
+            float sample = NonAtomicWorldgenRandom.sampleFloat(seed);
+            cir.setReturnValue(sample < probability ? this.replacementBlock : this.normalBlock);
         }
     }
 }