--- a/src/main/java/io/github/opencubicchunks/cubicchunks/mixin/core/common/chunk/MixinNoiseBasedChunkGenerator.java
+++ b/src/main/java/io/github/opencubicchunks/cubicchunks/mixin/core/common/chunk/MixinNoiseBasedChunkGenerator.java
@@ -150,7 +150,6 @@ public abstract class MixinNoiseBasedChunkGenerator {
                                 double density) {
         // optimization: we don't need to compute aquifer if we know that this block is already solid
         if (density > 0.0) {
-            ci.cancel();
             ci.setReturnValue(stoneSource.getBaseStone(x, y, z, this.settings.get()));
         }
     }