… src/main/java/io/github/opencubicchunks/cubicchunks/mixin/core/common/chunk/cave/MixinAquifer.java
… src/main/java/io/github/opencubicchunks/cubicchunks/mixin/core/common/chunk/cave/legacy/MixinCaveWorldCarver.java
… src/main/java/io/github/opencubicchunks/cubicchunks/mixin/core/common/chunk/cave/legacy/MixinWorldCarver.java
… src/main/java/io/github/opencubicchunks/cubicchunks/world/CubeWorldGenRandom.java
… src/main/resources/cubicchunks.mixins.core.json