This is quite a generally invasive PR addressing a wide range of targets. Most significant though are the optimizations to aquifers, which has been reduced from ~20% CPU time during generation to ~2%.

Loading a world with 12 threads:
- Pre-optimizations: 42.3s
- Post-optimizations: 37.5s

Loading a world with 2 threads:
- Pre-optimizations: 110.2s
- Post-optimizations: 80.9s